+++ R/package.R
@@ -3,8 +3,6 @@
 #' Numerical quadrature of functions of one variable over a finite
 #' or infinite interval with double exponential formulas.
 #'
-#' @docType package
-#' @name deformula
 #' @useDynLib deformula, .registration = TRUE
 #' @importFrom Rcpp sourceCpp
 "_PACKAGE"